chore(category): 카테고리-해시태그 연결 테이블 유사도 타입 변경

+++ apps/product-service/src/main/java/com/sub9/productservice/category/domain/entity/Category.java
@@ -64,7 +64,7 @@ public class Category extends BaseEntity {
                 .build();
     }
 
-    public CategoryHashtag linkManually(Hashtag hashtag, BigDecimal similarityScore) {
+    public CategoryHashtag linkManually(Hashtag hashtag, double similarityScore) {
         return CategoryHashtag.create(
                 this,
                 hashtag,
